restMainView 디테일 뷰 review ajax 수정 중

+++ src/main/java/com/always5/review/rest/service/RestService.java
@@ -15,5 +15,5 @@ public interface RestService {
 	public Dibs selectDibsForUserNo(Dibs dibsInfo);
 	public ArrayList<Review> selectReviewList(int restNo);
 	public ArrayList<Integer> selectRatingCount(int restNo);
-	public int operateRating(int restNo);
+	public double operateRating(int restNo);
 }
